martian: move errors detection code to separate file
